requirements.txt
streamlit>=1.37
streamlit-drawable-canvas>=0.9.3
numpy>=1.24
matplotlib>=3.7
Pillow>=10.0
